fix(meetings): make the agenda helper usable outside details
The redesigned composer surfaces "Help me write" from every section and
from Quick create, but `generateAgenda` hard-400'd unless meetingType,
title, and projectName were all present — so the helper was unusable
from exactly the places the redesign added it to. It now rejects only
when both the title and the free-text goal are absent, and the prompt
builder omits whichever descriptors are missing instead of inventing
them. The client guard was relaxed to match.

Three gaps the story asked about, addressed alongside:

- `maxCharacters` was destructured nowhere in the controller, so the
  client's agenda cap never reached the AI service. Now forwarded.
- There was no AI-specific rate limit — only the global 500/min /api
  limiter, nowhere near tight enough to bound LiteLLM calls. Adds
  `aiRateLimiter` (10/min) on POST /generate-agenda, keyed on the
  authenticated user with an IPv6-safe IP fallback for anonymous
  callers, so one user on a shared egress IP can't exhaust everyone's
  budget.
- Usage telemetry logged only the estimated duration. The success log
  now records which descriptors were supplied and the agenda length.

Refs #1464
