+++ packages/shared/src/interfaces/ai.interface.ts
@@ -8,13 +8,17 @@ import { AttendanceReconciliationCandidate, AttendanceReconciliationConfidence }
  * Request interface for AI agenda generation
  */
 export interface GenerateAgendaRequest {
-  /** Type of meeting for agenda generation */
-  meetingType: MeetingType;
-  /** Meeting title */
-  title: string;
-  /** Name of the project for contextualized agenda */
-  projectName: string;
-  /** Additional context or specific requirements */
+  /**
+   * Type of meeting for agenda generation.
+   * Optional: the composer can reach the helper before a type is chosen, in which case the prompt
+   * falls back to a generic project meeting.
+   */
+  meetingType?: MeetingType;
+  /** Meeting title. Optional — omitted from the prompt when the user hasn't set one yet. */
+  title?: string;
+  /** Name of the project for contextualized agenda. Optional, same reason as `title`. */
+  projectName?: string;
+  /** Additional context or specific requirements. Required when `title` is absent. */
   context?: string;
   /** Maximum characters allowed for the generated agenda */
   maxCharacters?: number;
